Send magic code to user in HTML instead of JSON
Lines 93 to 103 added **for users who are deploying the bot on a website**, if they want the user to receive the authentication code in html format (since some browsers don't support JSON..like IE) to improve user experience. In order to use this, simply comment line 91 and un-comment 93 to 103.

Thought it might help.